fizzbuzz now takes Range<u32> instead of Range<usize>

+++ src/fizzbuzz.rs
@@ -12,14 +12,14 @@ pub fn is_fizzbuzz(n: u32) -> bool {
     is_fizz(n) && is_buzz(n)
 }
 
-pub fn fizzbuzz(r: Range<usize>) -> String {
+pub fn fizzbuzz(r: Range<u32>) -> String {
     let mut output = String::new();
     for n in r.start..r.end {
         output = match n {
-            _ if is_fizzbuzz(n as u32) => output + "fizzbuzz\n",
-            _ if is_fizz(n as u32) => output + "fizz\n",
-            _ if is_buzz(n as u32) => output + "buzz\n",
-            _ => output + (n as u32).to_string().as_str() + "\n",
+            _ if is_fizzbuzz(n) => output + "fizzbuzz\n",
+            _ if is_fizz(n) => output + "fizz\n",
+            _ if is_buzz(n) => output + "buzz\n",
+            _ => output + (n).to_string().as_str() + "\n",
         }
     }
     output
